discount-service/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
requests
starlette
pydantic
python-dotenv
prometheus-fastapi-instrumentator
